[WIP] Typescriptify IrcHandler
